refactor: split event_handlers.rs unit tests into dedicated tests file
